Remove possible side effects of batch.

+++ src/main/java/de/photon/aacadditionpro/modules/checks/autofish/InhumanReactionPattern.java
@@ -38,13 +38,13 @@ class InhumanReactionPattern implements ListenerModule
                     // vl 6 is the maximum.
                     // Points = {{0, 1}, {8, 0}}
                     // Function: 1 - 0.125x
-                    for (byte b = 5; b > 0; b--) {
+                    for (int b = 5; b > 0; b--) {
                         if (user.getTimestampMap().recentlyUpdated(TimestampKey.LAST_FISH_BITE, (long) (1 - 0.125 * b) * fishingMilliseconds)) {
                             VerboseSender.getInstance().sendVerboseMessage("AutoFish-Verbose | Player " + user.getPlayer().getName() + " failed inhuman reaction");
 
                             // Flag for vl = b + 1 because there would otherwise be a "0-vl"
                             AutoFish.getInstance().getViolationLevelManagement().flag(event.getPlayer(),
-                                                                                      (int) b + 1,
+                                                                                      b + 1,
                                                                                       AutoFish.getInstance().getCancelVl(),
                                                                                       () -> event.setCancelled(true), () -> {});
                         }
